requirements.txt
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.12.1+cu113
torchvision==0.13.1+cu113
torchaudio==0.12.1
requests
opencv-python
openslide-python
h5py
pandas
scipy
scikit-learn
pyyaml
tensorboardX
protobuf
tqdm
matplotlib
scikit-image